pyproject.toml
[project]
name = "gr-fastapi-back-app"
version = "2.0.0"
description = "Sistema de Gestión de Reclamos - Backend API con FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI y servidor
    "fastapi[standard]>=0.120.2",
    "uvicorn[standard]>=0.38.0",
    
    # Base de datos PostgreSQL (mantener durante migración)
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",          # Driver PostgreSQL async
    "alembic>=1.13.0",          # Migraciones
    
    # Base de datos MongoDB (nuevo)
    "motor>=3.3.0",             # Driver MongoDB async oficial
    "beanie>=1.23.0",           # ODM para MongoDB (como SQLAlchemy)
    "pymongo[srv]>=4.6.0",      # Dependencia de Beanie
    
    # Validación y configuración
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    
    # Seguridad y autenticación
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    
    # Utilidades
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
